requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyzmq>=25.0.0
redis>=4.5.0
torch>=2.0.0
tensorflow>=2.12.0
lightgbm>=4.0.0
xgboost>=2.0.0
scikit-learn>=1.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3
requests>=2.31.0
aiofiles>=23.1.0
asyncio-contextmanager>=1.0.0
structlog>=23.1.0
python-json-logger>=2.0.7
MetaTrader5>=5.0.45
ccxt>=2.80.0
ta-lib>=0.4.28
joblib>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
